Moved formatting-related funcs/consts into common _output_formatting.py module
Details:

* Moved the following functions and constants from
  pywbemtools/pywbemcli/_common.py to a new module
  pywbemtools/_output_formatting.py:

  - output_format_is_table()
  - output_format_in_groups()
  - validate_output_format()
  - warning_msg()
  - display_text()
  - format_keys()
  - hide_empty_columns()
  - format_table()
  - fold_strings()
  - DEFAULT_MAX_CELL_WIDTH
  - TABLE_FORMATS
  - CIM_OBJECT_FORMATS
  - TEXT_FORMATS
  - DEFAULT_CIM_FORMAT
  - DEFAULT_TABLE_FORMAT
  - DEFAULT_TEXT_FORMAT
  - GROUPVALUES
  - OUTPUT_FORMAT_GROUPS
  - OUTPUT_FORMATS

* Moved unit testcases accordingly into a new test module.

* Cleaned up the implementation of output_format_in_groups() and cleaned up its
  interface by using None instead of an empty list for the valid_format_groups
  parameter to mean that any group is valid. There was no occurrence in the code
  that passed an empty list. Adjusted the unit test cases accordingly.
