+++ pywbemtools/pywbemcli/_cmd_profile.py
@@ -29,11 +29,11 @@ import click
 from pywbem import ValueMapping, Error
 
 from .pywbemcli import cli
-from ._common import format_table, pywbem_error_exception, \
-    validate_output_format
+from ._common import pywbem_error_exception
 from ._common_options import add_options, help_option
 from .._click_extensions import PywbemtoolsGroup, PywbemtoolsCommand, \
     CMD_OPTS_TXT, GENERAL_OPTS_TXT, SUBCMD_HELP_TXT
+from .._output_formatting import validate_output_format, format_table
 
 # NOTE: A number of the options use double-dash as the short form.  In those
 # cases, a third definition of the options without the double-dash defines
