+++ pywbemtools/pywbemcli/_cmd_server.py
@@ -32,13 +32,14 @@ from pywbem import Error, MOFCompiler
 from pywbem._mof_compiler import MOFWBEMConnection, MOFCompileError
 
 from .pywbemcli import cli
-from ._common import format_table, pywbem_error_exception, \
-    validate_output_format, display_text
+from ._common import pywbem_error_exception
 from ._common_options import add_options, help_option, namespace_option
 from .._click_extensions import PywbemtoolsGroup, PywbemtoolsCommand, \
     CMD_OPTS_TXT, GENERAL_OPTS_TXT, SUBCMD_HELP_TXT
 from ._cmd_namespace import cmd_namespace_list, cmd_namespace_interop
 from .._utils import pywbemcliwarn
+from .._output_formatting import validate_output_format, format_table, \
+    display_text
 
 # NOTE: A number of the options use double-dash as the short form.  In those
 # cases, a third definition of the options without the double-dash defines
