+++ pywbemtools/pywbemcli/_common.py
@@ -13,6 +13,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
 """
 Common Functions applicable across multiple components of pywbemcli
 """
@@ -21,13 +22,6 @@ from __future__ import absolute_import, print_function, unicode_literals
 
 import fnmatch
 import re
-from textwrap import fill
-from operator import itemgetter
-from collections import namedtuple
-try:
-    from collections import OrderedDict
-except ImportError:
-    from ordereddict import OrderedDict  # pylint: disable=import-error
 try:
     from collections.abc import Sequence
 except ImportError:
@@ -35,7 +29,6 @@ except ImportError:
 
 import six
 import click
-import tabulate
 from nocaselist import NocaseList
 from toposort import toposort_flatten
 
@@ -45,166 +38,7 @@ from pywbem import CIMInstanceName, CIMInstance, CIMClass, \
 from pywbem._nocasedict import NocaseDict
 
 from ._cimvalueformatter import cimvalue_to_fmtd_string
-
-
-DEFAULT_MAX_CELL_WIDTH = 100
-
-
-##############################################################
-#
-#     General option output format definitions and support for
-#     validating output formats
-#
-##############################################################
-
-# Definition of the format keywords within each format group
-TABLE_FORMATS = ('table', 'plain', 'simple', 'grid', 'psql', 'rst', 'html')
-CIM_OBJECT_FORMATS = ('mof', 'xml', 'repr', 'txt')
-TEXT_FORMATS = ('text',)
-
-# Definition of the default format for each format group
-DEFAULT_CIM_FORMAT = CIM_OBJECT_FORMATS[0]
-DEFAULT_TABLE_FORMAT = TABLE_FORMATS[2]
-DEFAULT_TEXT_FORMAT = TEXT_FORMATS[0]
-
-# Dictionary that relates group names to list of formats keywords for that
-# group and the default format for that group.  The order of the dictionary
-# entry determines the priority of selecting formats.
-GROUPVALUES = namedtuple('GROUPVALUES', 'keywords default')
-OUTPUT_FORMAT_GROUPS = OrderedDict(
-    [('CIM', GROUPVALUES(CIM_OBJECT_FORMATS, DEFAULT_CIM_FORMAT)),
-     ('TABLE', GROUPVALUES(TABLE_FORMATS, DEFAULT_TABLE_FORMAT)),
-     ('TEXT', GROUPVALUES(TEXT_FORMATS, DEFAULT_TEXT_FORMAT))])
-
-# All of the format strings
-OUTPUT_FORMATS = tuple(item for fmts in OUTPUT_FORMAT_GROUPS.values()
-                       for item in fmts.keywords)
-
-
-def output_format_is_table(output_format):
-    """
-    Return True if output format is a table form.
-
-    Parameters:
-
-      output_format (:term:`string`):
-         String containing the output_format keyvalue.
-
-    Returns:
-      True: if `output_format` is one of the Table format keyword choises
-      False: if `output_format` is not a Table format.
-    """
-    return output_format in TABLE_FORMATS
-
-
-def output_format_in_groups(output_format, format_groups):
-    """
-    Return True if output format is in the group defined by format_group
-
-    Parameters:
-
-      output_format (:term:`string`):
-         String containing the output_format keyvalue.
-
-      output_groups (:term:`string` or list of :term:`string`):
-         String containing the format group ot list of strings representing
-         multiple groups. This must be one of the keys in the table
-         OUTPUT_FORMAT_GROUPS.  If the
-
-    Returns:
-      True: if it is in the group or one of the list of groups
-      False: if it is not in any of the the groups.
-
-    Raises:
-        KeyError: If format_group parameter is not a valid group
-    """
-    if isinstance(format_groups, (list, tuple)):
-        for group in format_groups:
-            if output_format_in_groups(output_format, group):
-                return True
-        return False
-
-    # test if output_format string is in a group
-    return output_format in OUTPUT_FORMAT_GROUPS[format_groups]
-
-
-def validate_output_format(output_format, valid_format_groups,
-                           default_format=None):
-    """
-    Tests for valid format groups and provides a default format if the
-    context.output_format is None.
-
-    Parameters:
-
-      output_format (:term:`string` or None):
-        The output format string to be validated (normally provided by input
-        (i.e. ContextObj)) or None if there is no defined format for this
-        command execution.
-
-      valid_format_groups (list of :term:`string` or :term:`string`):
-        One or more strings in a list where the allowed strings are: 'TABLE',
-        'CIM', or 'TEXT'. An empty list implies any output group if valid. A
-        single string may be used to designate a single group
-
-      default_format (:term:`string` or None):
-        One of the valid format definitions or None.  The format must be in the
-        OUTPUT_FORMATS list. If None, the default format for the first group in
-        `valid_format_groups` is returned. This format keyword is returned if
-        `output_format` is None.
-
-    Returns:
-       :term:`string` containing output format keyword to be used.
-
-    Raises:
-        click.ClickException if format invalid for command
-    """
-
-    # These are asserts because they are coding errors in the input parameters
-    # for this function
-    if isinstance(valid_format_groups, six.string_types):
-        valid_format_groups = [valid_format_groups]
-
-    # Code error if this assert fails
-    assert all(element in OUTPUT_FORMAT_GROUPS for element in
-               valid_format_groups)
-
-    if default_format:
-        assert default_format in OUTPUT_FORMATS
-
-    # Confirm default matches group.
-    # CIM object type has priority over Table
-    if output_format:
-        # Do invalid message here but really should be assert because this
-        # is a error in the data defined for the validate call.
-        for group, value in OUTPUT_FORMAT_GROUPS.items():
-            if group in valid_format_groups:
-                if output_format in value.keywords:
-                    return output_format
-        if not valid_format_groups:
-            return output_format
-
-    else:   # output_format is None
-        if default_format:
-            return default_format
-        if valid_format_groups:
-            for group, value in OUTPUT_FORMAT_GROUPS.items():
-                if valid_format_groups[0] == group:
-                    return value.default
-            return OUTPUT_FORMAT_GROUPS['CIM'].default
-        return OUTPUT_FORMAT_GROUPS['CIM'].default
-
-    valid_formats = ""
-    for name, value in OUTPUT_FORMAT_GROUPS.items():
-        fmt_list = value.keywords
-        if name in valid_format_groups:
-            if valid_formats:
-                valid_formats += "; "
-            valid_formats += '{0} formats: "({1})"'.format(
-                name, '", "'.join(fmt_list))
-
-    raise click.ClickException('Output format "{}" not allowed for this '
-                               'command. Only {} allowed.'.
-                               format(output_format, valid_formats))
+from .._output_formatting import DEFAULT_MAX_CELL_WIDTH
 
 
 ######################################################################
@@ -253,11 +87,6 @@ def resolve_propertylist(propertylist):
         propertylist = pl
 
     return propertylist
-
-
-def warning_msg(msg):
-    """Issue the msg param as warning prefixed by WARNING: to stderr"""
-    click.echo('WARNING: {}'.format(msg), err=True)
 
 
 ######################################################################
@@ -1042,14 +871,6 @@ def get_subclass_names(classes, classname=None, deep_inheritance=None):
     return NocaseList(rtn_classnames)
 
 
-def display_text(text, output_format=None):  # pylint: disable=unused-argument
-    """
-    Display the text output format. Currently this simply outputs to
-    click.echo
-    """
-    click.echo(text)
-
-
 def shorten_path_str(path, replacements, fullpath):
     """
     Create a short-form path str from the input CIMInstanceName with selected
@@ -1102,305 +923,6 @@ def shorten_path_str(path, replacements, fullpath):
             name_str = name_str.replace("{}={}".format(key, magicvalue), "~", 1)
 
     return name_str
-
-
-def format_keys(obj, max_width):
-    """
-    Format the keys of a dictionary of keybindings as text for display. Formats
-    multiple keybindings on each line within the max_width
-
-    Parameters:
-
-      obj (:class:`pwbem.CIMInstanceName`):
-        Instance name from which keybindings are to be extracted for
-        formatting.
-
-    Returns:
-        :term:`string` containing the keys from the input obj formatted for
-        display at within the defined width.
-    """
-    def get_wbemurikeys(obj):
-        """
-        Create wbem_uri from CIMInstanceName and separate out key component
-        for return.
-        """
-        wbem_uri = obj.to_wbem_uri()
-        wbem_uri_keys = wbem_uri[wbem_uri.find('.'):]
-        wbem_uri_keys = wbem_uri_keys[1:]
-        return wbem_uri_keys
-
-    assert isinstance(obj, CIMInstanceName)
-    # clear the host and namespace
-    myobj = obj.copy()
-    myobj.host = None
-    myobj.namespace = None
-    wbem_uri_keys = get_wbemurikeys(myobj)
-
-    # Too long for width. Fold the keys on multiple lines
-    if len(wbem_uri_keys) > max_width:
-        wbem_uri_keys = ''
-        line_len = 0
-        for key, value in myobj.keybindings.items():
-            one_key_obj = get_wbemurikeys((CIMInstanceName('x', {key: value})))
-            if wbem_uri_keys:
-                if line_len + len(one_key_obj) > max_width:
-                    wbem_uri_keys += '\n{}'.format(one_key_obj)
-                    line_len = 0
-                else:
-                    wbem_uri_keys += ',{}'.format(one_key_obj)
-                    line_len += len(one_key_obj) + 1
-
-            else:  # must put on first line even if too long
-                wbem_uri_keys += one_key_obj
-                line_len = len(one_key_obj) + 1
-
-    return wbem_uri_keys
-
-
-def hide_empty_columns(headers, rows):
-    """
-    Removes columns from rows if the colmuns are considered empty.
-    The definiton of an empty row is:
-    1. All entries for the column in all rows are None or "" if type string.
-    2. All entries for the column in all rows are None if number.
-
-    Parameters:
-      headers (list of :term:`string`)
-        The strings that represent the column titles of an array of rows.
-
-      rows (list of list of TBD):
-        The rows of a table where each row is a list of the items that
-        represent the columns of the row.
-
-    Returns new rows and headers
-    """
-    def column_is_empty(rows, column):
-        """
-        Determine if entries for defined column in all rows are considered
-        empty.
-        Returns True if all are empty. Otherwise returns False
-        """
-        for row in rows:
-            if isinstance(row[column], six.integer_types) and \
-                    row[column] is not None:
-                return False
-            if row[column]:
-                return False
-        return True
-
-    # Remove empty rows
-    len_hdr = len(headers)
-    for row in rows:
-        assert len(row) == len_hdr, "row: {}\nhdrs: {}". \
-            format(row, headers)
-    for column in range(len(headers) - 1, -1, -1):
-        if column_is_empty(rows, column):
-            if isinstance(headers, tuple):
-                headersl = list(headers)
-                del headersl[column]
-                headers = tuple(headersl)
-            else:
-                del headers[column]
-            for row in rows:
-                del row[column]
-
-    return headers, rows
-
-
-def format_table(rows, headers, title=None, table_format='simple',
-                 sort_columns=None, hide_empty_cols=None, float_fmt=None):
-    """
-    General print table function.  Prints a list of lists in a
-    table format where each inner list is a row.
-    This code is temporary while the tabulate package is updated
-
-    Parameters:
-
-      headers (list strings):
-        where each string is a table column name or None if no header is to be
-        attached
-
-      table_data (list of lists):
-        where each the top level iterables represents the list of rows
-        and each row is an iterable of strings for the data in that row.
-
-      title (:term:`string`):
-         Optional title to be places io the output above the table.
-         No title is output if this parameter is None
-
-      table_format (:term:`string`):
-        Output format defined by the string and limited to one of the
-        choice of table formats defined in TABLE_FORMATS list
-
-      output_file (:term:`string`):
-        If not None, a file name to which the output formatted data is sent.
-
-      sort_columns (int or list of int that defines sort):
-        Defines the cols that will be sorted. If int, it defines the column
-        that will be sorted. If list of int, the sort is in sort order of
-        cols in the list (i.e. minor sorts to the left, major sorts to the
-        right). Note that entries in each row of the columns to be sorted
-        must be of the same type (int, str, etc.) to be sortable.
-
-    hide_empty_cols (:class:`py:bool`):
-        If this flag is True any columns that are completely blank are
-        hiddend and the column header is removed from the headers.
-        Uses the function hide_empty_columns
-
-    float_fmt (:term:`string` or list of :term:`string`):
-        String defining the formating of floating point format, either
-        universal for all floating point if a single string or for selected
-        columns if a list is provided.
-        The string defines the number of decimals after the decimal point as
-        ".<number>f". Thus ".4f" tells the formatter to use 4 digits after the
-        decimal point
-
-    Returns:
-        :term:`string`: Returns the formatted table as a string
-
-    Raises:
-        click.ClickException if invalid table format string
-    """
-    if hide_empty_cols:
-        headers, rows = hide_empty_columns(headers, rows)
-    if sort_columns is not None:
-        if isinstance(sort_columns, int):
-            rows = sorted(rows, key=itemgetter(sort_columns))
-        elif isinstance(sort_columns, (list, tuple)):
-            rows = sorted(rows, key=itemgetter(*sort_columns))
-        else:
-            assert False, "Sort_columns must be int or list/tuple of int"
-
-    if table_format is None:
-        table_format = 'table'
-    if table_format == 'table':
-        table_format = 'psql'
-    if not output_format_is_table(table_format):
-        raise click.ClickException('Invalid table format {}.'
-                                   .format(table_format))
-
-    # Required because tabulate applies an internally defined format
-    # when floatfmt not defined.
-    if float_fmt:
-        result = tabulate.tabulate(rows, headers, tablefmt=table_format,
-                                   floatfmt=float_fmt)
-    else:
-        result = tabulate.tabulate(rows, headers, tablefmt=table_format)
-
-    if title:
-        if table_format == 'html':
-            # Insert caption element immediatly after table
-            assert result.startswith("<table>")
-            replacement = "<table>\n<caption>{0}</caption>".format(title)
-            result = result.replace("<table>", replacement, 1)
-        else:
-            result = '{0}\n{1}'.format(title, result)
-    return result
-
-
-def fold_strings(input_strings, max_width, break_long_words=False,
-                 break_on_hyphens=False, fold_list_items=False, separator=', ',
-                 initial_indent='', subsequent_indent=''):
-    """
-    Fold a string or a list/tuple of strings within a maximum width and return
-    a folded string that fits within the width defined by max_width. If input
-    is a list of strings, the fold_list_itmes defines whether they create
-    separate lines in the output or are concatenated into a single string
-    before folding the string. This implementation refolds strings that already
-    contain EOL characters and removes any existing folds that do not match the
-    max_width criteria.  Lists of strings may be folded with one string per
-    line or concatenated and then folded.
-
-    Parameters:
-
-      input_strings (:term:`string` or list of :term:`string`):
-        The string that will be contents of into the cell. This string may
-        already include multiple lines.
-
-      max_width (:term:`integer`):
-        Maximum width of cell containing the resulting string.  Data is
-        folded into multiple lines to fit into this width.
-
-      break_long_words (:class:`py:bool`):
-        Boolean that forces long words to be broken if True
-        If False, long words will not break at the max width
-
-      break_on_hyphens (:class:`py:bool`):
-        If True use hypens as word separator.
-
-      fold_list_items (:class:`py:bool`):
-        If True, force fold for each item in list/tupe of input strings if
-        single line concatenation is longer than max_width. Otherwise list is
-        contatenated and the result folded.
-
-      separator (:term:`string`):
-        String that is separator between list items when folded into
-        string.
-
-      initial_indent (:term:`integer` or :term:`string`):
-        Integer  or string defining the number of characters of indent for
-        the first line if single string or the first line of each string in
-        the list if input_strings is a list and fold_list_items is True.
-
-      subsequent_indent (:term:`integer`):
-        Integer or string, defining the number of characters of indent for
-        the all but the first lineif single string or the first line of
-        each string in the list if input_strings is a list and
-        fold_list_items is True.
-
-    Returns:
-        String representing the folded input_strings
-    """
-    def indent_str(indent):
-        """
-        Maps indent to string of indent characters if necessary.
-        If None, return None
-        If integer, map to indent characters
-        """
-        if indent is None:
-            return ""
-        if isinstance(indent, six.string_types):
-            return indent
-        return ' ' * indent
-
-    initial_indent = indent_str(initial_indent)
-    subsequent_indent = indent_str(subsequent_indent)
-
-    if isinstance(input_strings, (list, tuple)):
-        if separator is None:
-            separator = ', '
-        build_str = separator.join(input_strings)
-        # Return contatentated list if within cell width
-        if len(build_str) <= max_width:
-            return build_str
-
-        if fold_list_items:
-            folded_strings = []
-            for str_item in input_strings:
-                folded_strings.append(
-                    fold_strings(str_item, max_width,
-                                 break_long_words=break_long_words,
-                                 break_on_hyphens=break_on_hyphens,
-                                 initial_indent=initial_indent,
-                                 subsequent_indent=subsequent_indent))
-            return "\n".join(folded_strings)
-        input_strings = build_str
-
-    # process single string
-    input_string = input_strings
-    assert isinstance(input_string, six.string_types)
-
-    if len(input_string) <= max_width:
-        return input_string
-
-    # use textwrap fill to fold the string
-    folded_string = fill(input_string, max_width,
-                         break_long_words=break_long_words,
-                         break_on_hyphens=break_on_hyphens,
-                         initial_indent=initial_indent,
-                         subsequent_indent=subsequent_indent)
-
-    return folded_string
 
 
 def pywbem_error_exception(exc, intro=None):
