+++ pywbemtools/pywbemcli/_display_cimobjects.py
@@ -31,10 +31,11 @@ from pywbem import CIMInstanceName, CIMInstance, CIMClass, \
     CIMError, CIM_ERR_NOT_SUPPORTED
 from pywbem._nocasedict import NocaseDict
 
-from ._common import format_table, fold_strings, DEFAULT_MAX_CELL_WIDTH, \
-    output_format_is_table, sort_cimobjects, format_keys
+from ._common import sort_cimobjects
 from ._cimvalueformatter import cimvalue_to_fmtd_string
 from .._utils import get_terminal_width
+from .._output_formatting import DEFAULT_MAX_CELL_WIDTH, \
+    output_format_is_table, format_table, fold_strings, format_keys
 
 INT_TYPE_PATTERN = re.compile(r'^[su]int(8|16|32|64)$')
